페이징, 검색 제외 완성

+++ src/main/java/com/ldg/blog/service/CommentService.java
@@ -1,11 +1,16 @@
 package com.ldg.blog.service;
 
+import java.util.List;
+
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.ldg.blog.model.ReturnCode;
 import com.ldg.blog.model.comment.dto.ReqDetailDto;
 import com.ldg.blog.model.comment.dto.RespDetailDto;
+import com.ldg.blog.model.user.User;
 import com.ldg.blog.repository.CommentRepository;
 
 @Service
@@ -14,11 +19,21 @@ public class CommentService {
 	@Autowired
 	private CommentRepository commentRepository;
 	
+	@Autowired
+	private HttpSession session;
+	
+	
+	public List<RespDetailDto> 목록보기(int id) {
+		
+		return commentRepository.findByPostId(id);
+		
+	}
+	
 	
 	public RespDetailDto 댓글쓰기(ReqDetailDto dto) {
 		int result = commentRepository.save(dto);
 		
-		System.out.println(dto.getId());
+		
 		
 		if(result == 1) {
 			RespDetailDto respDetailDto = commentRepository.findById(dto.getId());
@@ -33,7 +48,20 @@ public class CommentService {
 	}
 	
 	public int 댓글삭제(int id) {
-		return commentRepository.delete(id);			
+		//해당 댓글은 누가 씀??
+		
+		RespDetailDto comment = commentRepository.findById(id);
+		
+		//지금 로그인 주체는 누구인가?
+		User principal = (User) session.getAttribute("principal");
+		
+		
+		if(comment.getUsername().equals(principal.getUsername())) {
+			return commentRepository.delete(id);		
+		}else {
+			return ReturnCode.권한없음;
+		}		
+			
 		
 	}
